Melhora OpenAiService com configuração externalizada e melhor tratamento de erros
