+++ src/main/java/br/com/cesurgmarau/trabalho_final/infra/repository/ComentarioRepositoryImpl.java
@@ -10,15 +10,14 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 import java.util.Optional;
 
-
 @Repository
-@Transactional
 public class ComentarioRepositoryImpl implements ComentarioRepository {
 
     @PersistenceContext
     private EntityManager entityManager;
 
     @Override
+    @Transactional
     public Comentario salvar(Comentario comentario) {
         entityManager.persist(comentario);
         return comentario;
@@ -26,45 +25,42 @@ public class ComentarioRepositoryImpl implements ComentarioRepository {
 
     @Override
     public Optional<Comentario> buscarPorId(Integer id) {
-        Comentario comentario = entityManager.find(Comentario.class, id);
-        return Optional.ofNullable(comentario);
+        return Optional.ofNullable(entityManager.find(Comentario.class, id));
     }
 
     @Override
     public List<Comentario> buscarTodos() {
-        return entityManager.createQuery(
-                "SELECT c FROM Comentario c", Comentario.class
-        ).getResultList();
+        return entityManager
+                .createQuery("SELECT c FROM Comentario c", Comentario.class)
+                .getResultList();
     }
 
     @Override
     public List<Comentario> buscarPorProdutoId(Integer produtoId) {
-        return entityManager.createQuery(
-                        "SELECT c FROM Comentario c WHERE c.produtoId = :produtoId", Comentario.class
-                )
+        return entityManager
+                .createQuery("SELECT c FROM Comentario c WHERE c.produtoId = :produtoId", Comentario.class)
                 .setParameter("produtoId", produtoId)
                 .getResultList();
     }
 
     @Override
     public List<Comentario> buscarPorUsuarioId(Integer usuarioId) {
-        return entityManager.createQuery(
-                        "SELECT c FROM Comentario c WHERE c.usuarioId = :usuarioId", Comentario.class
-                )
+        return entityManager
+                .createQuery("SELECT c FROM Comentario c WHERE c.usuarioId = :usuarioId", Comentario.class)
                 .setParameter("usuarioId", usuarioId)
                 .getResultList();
     }
 
     @Override
     public List<Comentario> buscarPorSentimento(String sentimento) {
-        return entityManager.createQuery(
-                        "SELECT c FROM Comentario c WHERE c.sentimento = :sentimento", Comentario.class
-                )
+        return entityManager
+                .createQuery("SELECT c FROM Comentario c WHERE c.sentimento = :sentimento", Comentario.class)
                 .setParameter("sentimento", sentimento)
                 .getResultList();
     }
 
     @Override
+    @Transactional
     public void deletar(Integer id) {
         Comentario comentario = entityManager.find(Comentario.class, id);
         if (comentario != null) {
